+++ src/FITSHeaders.jl
@@ -31,6 +31,10 @@ export
     FITS_BLOCK_SIZE,
     FITS_SHORT_KEYWORD_SIZE
 
+@static if VERSION ≥ v"1.11"
+    include("public.jl")
+end
+
 using Requires
 
 # Enumeration of keyword value type identifiers.