+++ src/public.jl
@@ -1,0 +1,4 @@
+public is_structural
+public is_comment
+public is_naxis
+public is_end